Add files via upload
i modified Salveo Magprobe to use two 8X12X3.5 bearings and 3 m2.5X20mm screws, the Z endstop is mounted to the Y carrage like a 2.4. You will need a ender bed on the standard switchwire build for this to work, also you will need a new y carrage for this to fit  and work. with this setup i get to use auto-z  with the klicky macros.